api/requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
click==8.1.8
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-pagination==0.12.34
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.0
pygments==2.18.0
PyMySQL==1.1.1
pytest==8.3.4
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
typer==0.15.1
typing-extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
